+++ test/test_mapping.cpp
@@ -80,7 +80,7 @@ int main() {
     realMapTotal[i].resize(ids.size());
     realMapDetailed[i].resize(ids.size());
     for (size_t j = 0; j < ids.size(); ++j) {
-      realMap[i][j] = result->getSubstitutionCount(ids[j]);
+      realMap[i][j] = static_cast<double>(result->getSubstitutionCount(ids[j]));
       realMapTotal[i][j].resize(totReg->getNumberOfSubstitutionTypes());
       realMapDetailed[i][j].resize(detReg->getNumberOfSubstitutionTypes());
       result->getSubstitutionCount(ids[j], *totReg, realMapTotal[i][j]);
